requirements.txt
opencv-python
Pillow
pyzbar
pytesseract
easyocr
pandas
torch
scikit-learn
streamlit
tensorflow
opencv-python
Pillow
pytesseract
pymupdf
Faker